Change StackCount type to int64
Blueprint only accepts signed integer types. Changing the type from
uint32 to int64 preserves the value range while improving the
interoperability with the Blueprint scripting.

+++ Plugins/GameSystems/InventorySystem/Source/InventorySystem/Private/InventorySystem.cpp
@@ -3,4 +3,4 @@
 #include "InventorySystem.h"
 
 // ReSharper disable once CppVariableCanBeMadeConstexpr
-const uint32 InventorySystem::GMaxStackCountDefault = 1;
+const int64 InventorySystem::GMaxStackCountDefault = 1;
